+++ src/main/java/com/nhnacademy/hexashoppingmallservice/controller/review/ReviewController.java
@@ -3,6 +3,8 @@ package com.nhnacademy.hexashoppingmallservice.controller.review;
 import com.nhnacademy.hexashoppingmallservice.dto.book.ReviewRequestDTO;
 import com.nhnacademy.hexashoppingmallservice.projection.review.ReviewProjection;
 import com.nhnacademy.hexashoppingmallservice.service.review.ReviewService;
+import com.nhnacademy.hexashoppingmallservice.util.JwtUtils;
+import jakarta.servlet.http.HttpServletRequest;
 import jakarta.validation.Valid;
 import lombok.RequiredArgsConstructor;
 import org.bouncycastle.pqc.jcajce.provider.lms.LMSKeyFactorySpi;
@@ -19,6 +21,7 @@ import java.util.List;
 @RequiredArgsConstructor
 public class ReviewController {
     private final ReviewService reviewService;
+    private final JwtUtils jwtUtils;
 
     /**
      * 새로운 리뷰를 생성하는 엔드포인트
@@ -28,11 +31,12 @@ public class ReviewController {
      * @param bookId           리뷰를 작성할 책의 ID
      * @return 생성된 리뷰의 상태 코드
      */
-    @PostMapping("/auth/members/{memberId}/books/{bookId}/reviews")
+    @PostMapping("/members/{memberId}/books/{bookId}/reviews")
     public ResponseEntity<Void> createReview(
             @Valid @RequestBody ReviewRequestDTO reviewRequestDTO,
             @PathVariable String memberId,
-            @PathVariable Long bookId) {
+            @PathVariable Long bookId, HttpServletRequest request) {
+        jwtUtils.ensureUserAccess(request, memberId);
         reviewService.createReview(reviewRequestDTO, memberId, bookId);
         return new ResponseEntity<>(HttpStatus.CREATED);
     }
@@ -44,10 +48,11 @@ public class ReviewController {
      * @param pageable 페이징 및 정렬 정보
      * @return 회원의 리뷰 목록 페이지
      */
-    @GetMapping("/auth/members/{memberId}/reviews")
+    @GetMapping("/members/{memberId}/reviews")
     public ResponseEntity<List<ReviewProjection>> getReviewsFromMember(
             @PathVariable String memberId,
-            Pageable pageable) {
+            Pageable pageable, HttpServletRequest request) {
+        jwtUtils.ensureUserAccess(request, memberId);
         List<ReviewProjection> reviews = reviewService.getReviewsFromMember(pageable, memberId);
         return ResponseEntity.ok(reviews);
     }
@@ -59,15 +64,18 @@ public class ReviewController {
      * @param reviewId         수정할 리뷰의 ID
      * @return 수정된 리뷰의 상태 코드
      */
-    @PutMapping("/auth/reviews/{reviewId}")
+    @PutMapping("/reviews/{reviewId}")
     public ResponseEntity<Void> updateReview(
             @Valid @RequestBody ReviewRequestDTO reviewRequestDTO,
-            @PathVariable Long reviewId) {
+            @PathVariable Long reviewId, HttpServletRequest request) {
+        String token = jwtUtils.getTokenFromRequest(request);
+        String memberId = jwtUtils.getUsernameFromToken(token);
+        jwtUtils.ensureUserAccess(request, memberId);
         reviewService.updateReview(reviewRequestDTO, reviewId);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    @GetMapping("/auth/books/{bookId}/reviews")
+    @GetMapping("/books/{bookId}/reviews")
     public ResponseEntity<List<ReviewProjection>> getReviewsFromBook(@PathVariable Long bookId, Pageable pageable) {
         List<ReviewProjection> reviews = reviewService.getReviewsFromBook(pageable, bookId);
         return ResponseEntity.ok(reviews);
@@ -79,19 +87,25 @@ public class ReviewController {
      * @param reviewId 삭제할 리뷰의 ID
      * @return 삭제된 리뷰의 상태 코드
      */
-    @DeleteMapping("/auth/reviews/{reviewId}")
-    public ResponseEntity<Void> deleteReview(@PathVariable Long reviewId) {
+    @DeleteMapping("/reviews/{reviewId}")
+    public ResponseEntity<Void> deleteReview(@PathVariable Long reviewId, HttpServletRequest request) {
+        String token = jwtUtils.getTokenFromRequest(request);
+        String memberId = jwtUtils.getUsernameFromToken(token);
+        jwtUtils.ensureUserAccess(request, memberId);
         reviewService.deleteReview(reviewId);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
 
-    @GetMapping("/admin/reviews")
+    @GetMapping("/reviews")
     public ResponseEntity<List<ReviewProjection>> getReviewsFromAdmin(Pageable pageable) {
         return ResponseEntity.ok(reviewService.getReviewsIsBlocked(pageable));
     }
 
-    @PutMapping("/reviews/{reviewId}")
-    public ResponseEntity<Void> updateReviewBlock(@PathVariable Long reviewId, @RequestParam boolean block) {
+    @PutMapping("/reviews/{reviewId}/block")
+    public ResponseEntity<Void> updateReviewBlock(@PathVariable Long reviewId, @RequestParam boolean block, HttpServletRequest request) {
+        String token = jwtUtils.getTokenFromRequest(request);
+        String memberId = jwtUtils.getUsernameFromToken(token);
+        jwtUtils.ensureUserAccess(request, memberId);
         reviewService.updateBlocked(reviewId, block);
         return new ResponseEntity<>(HttpStatus.NO_CONTENT);
     }
